requirements.txt
flask==2.3.3
pyautogen==0.8.5
openai==1.72.0
python-dotenv==1.1.0
flaml==2.3.4
tiktoken==0.9.0
google-generativeai==0.8.4 